+++ app/conf/development/settings.py
@@ -72,10 +72,23 @@ DEFAULT_FROM_EMAIL = 'test@example.com'
 
 DATABASES = {
     'default': {
+        'ENGINE': 'django.db.backends.postgresql',
+        'NAME': 'blackjack001',
+        'USER': 'eddy',
+        'PASSWORD': 'vanhalen',
+        'HOST': 'localhost',
+        'PORT': '5432',
+    }
+}
+
+"""
+DATABASES = {
+    'default': {
         'ENGINE': 'django.db.backends.sqlite3',
         'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
     }
 }
+"""
 
 AUTH_PASSWORD_VALIDATORS = [
     {