+++ app/settings.py
@@ -1,6 +1,10 @@
 import os
+import sys
 
+"""
 IS_PRODUCTION = os.environ.get('IS_PRODUCTION')
+"""
+IS_PRODUCTION = (sys.argv[1] != 'runserver')
 
 if IS_PRODUCTION:
     from .conf.production.settings import *